finished js et csrf hell. Re-facto wip

+++ app/src/MVC/controllers/PictureController.class.php
@@ -14,11 +14,12 @@ class PictureController extends Controller
 	
 	public function workshop($params)
 	{
-		$user = $this->container->get_auth()->being_auth(true);
+		$user = $this->auth->being_auth(true);
 		
 		$data = [
 			'title' => 'Workshop',
-			'user' => $user
+			'user' => $user,
+			'csrf' => $this->form_key->outputKey()
 		];
 		$this->container->get_View("workshop.php", $data);
 	}
@@ -26,17 +27,29 @@ class PictureController extends Controller
 	public function save($params)
 	{
 		// $params -> base64encode(str) and frames(array of objects)
-		$user = $this->container->get_auth()->being_auth(true);
+		$user = $this->auth->being_auth(true, true);
+		$no_csrf = $this->security->check_csrf('osef', true);
+		$no_validation = $this->security->validate_inputs_format($params, 'osef', true);
+		$form_key = $this->form_key->get_key();
+		if ($this->security->ajax_secure_and_display($user, $no_validation, $no_csrf, $form_key))
+			exit;
+		
 		$img = $this->image_helper->merge_image($params);
 		$img_path = $this->image_helper->save_image($img, $user->get_id());
 		$picture_id = $this->container->get_PictureCollection()->new(['user_id' => $user->get_id(), 'path' => $img_path]);
-		echo json_encode(['src' => $img_path, 'picture_id' => "pic".$picture_id]);
+		echo json_encode(['src' => $img_path, 'picture_id' => "pic".$picture_id, 'csrf' => $form_key]);
 	}
 
 	public function delete($params)
 	{
 		// $params -> picture_id
-		$user = $this->container->get_auth()->being_auth(true);
+		$user = $this->auth->being_auth(true, true);
+		$no_csrf = $this->security->check_csrf('osef', true);
+		$no_validation = $this->security->validate_inputs_format($params, 'osef', true);
+		$form_key = $this->form_key->get_key();
+		if ($this->security->ajax_secure_and_display($user, $no_validation, $no_csrf, $form_key))
+			exit;
+
 		$picture = $this->container->get_PictureCollection()->find('id', $params['picture_id']);
 		if ($picture->get_user_id() !== $user->get_id())
 		{
@@ -45,35 +58,53 @@ class PictureController extends Controller
 		}
 		$picture->delete();
 		$this->image_helper->delete_image($picture->get_path());
-		echo $params['picture_id'];
+		echo json_encode(['picture_id' => $params['picture_id'], 'csrf' => $form_key]);
 	}
 
 	public function like($params)
 	{
 		// $params -> picture_id
-		$user = $this->container->get_auth()->being_auth(true);
+		$user = $this->auth->being_auth(true, true);
+		$no_csrf = $this->security->check_csrf('osef', true);
+		$no_validation = $this->security->validate_inputs_format($params, 'osef', true);
+		$form_key = $this->form_key->get_key();
+		if ($this->security->ajax_secure_and_display($user, $no_validation, $no_csrf, $form_key))
+			exit;
+
 		$picture = $this->container->get_PictureCollection()->find('id', $params['picture_id']);
 		$like_id = $this->container->get_LikeCollection()->new(['user_id' => $user->get_id(), 'picture_id' => $params['picture_id']]);
-		echo $like_id;
+		echo json_encode(['like_id' => $like_id, 'csrf' => $this->form_key->get_key()]);
 	}
 
 	public function dislike($params)
 	{
 		// $params -> like_id
-		$user = $this->container->get_auth()->being_auth(true);
+		$user = $this->auth->being_auth(true, true);
+		$no_csrf = $this->security->check_csrf('osef', true);
+		$no_validation = $this->security->validate_inputs_format($params, 'osef', true);
+		$form_key = $this->form_key->get_key();
+		if ($this->security->ajax_secure_and_display($user, $no_validation, $no_csrf, $form_key))
+			exit;
+
 		$like = $this->container->get_LikeCollection()->find('id', $params['like_id']);
 		$picture_id = $like->get_picture_id();
 		$like->delete();
-		echo 'pic'.$picture_id;
+		echo json_encode(['picture_id' => 'pic'.$picture_id, 'csrf' => $this->form_key->get_key()]);
 		
 	}
 
 	public function comment($params)
 	{
 		// $params -> picture_id and comment
-		$user = $this->container->get_auth()->being_auth(true);
+		$user = $this->auth->being_auth(true, true);
+		$no_csrf = $this->security->check_csrf('osef', true);
+		$no_validation = $this->security->validate_inputs_format($params, 'osef', true);
+		$form_key = $this->form_key->get_key();
+		if ($this->security->ajax_secure_and_display($user, $no_validation, $no_csrf, $form_key))
+			exit;
+
 		$picture = $this->container->get_PictureCollection()->find('id', $params['picture_id']);
 		$comment_id = $this->container->get_CommentCollection()->new(['user_id' => $user->get_id(), 'picture_id' => $params['picture_id'], 'comment' => $params['comment']]);
-		echo json_encode(['comment' => $params['comment'], 'owner_profile' => $user->get_gravatar_hash(), 'owner_login' => $user->get_login()]);
+		echo json_encode(['comment' => $params['comment'], 'owner_profile' => $user->get_gravatar_hash(), 'owner_login' => $user->get_login(), 'csrf' => $form_key]);
 	}
 }
